Restructure for readability
